api logic will follow a modular structure
