+++ server/src/api/mod.rs
@@ -1,0 +1,5 @@
+mod v1;
+
+pub fn get_v1_routes() -> Vec<rocket::Route> {
+    v1::get_all_routes()
+}
